third_party/WebKit/LayoutTests/http/tests/inspector/bindings/shadowdom-navigator-expected.txt

Verify that navigator shows proper UISourceCodes as shadow dom styles and scripts are added and removed.


Running: dumpInitialNavigator
top
  127.0.0.1:8000
    inspector
      bindings
        shadowdom-navigator.html
        bindings-test.js
      inspector-test.js

Running: attachShadow1
top
  127.0.0.1:8000
    inspector
      bindings
        resources
          sourcemap-typescript.ts
          sourcemap-sass.scss
        shadowdom-navigator.html
        bindings-test.js
      inspector-test.js
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: attachShadow2
top
  127.0.0.1:8000
    inspector
      bindings
        resources
          sourcemap-typescript.ts
          sourcemap-sass.scss
        shadowdom-navigator.html
        bindings-test.js
      inspector-test.js
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: detachShadow1
top
  127.0.0.1:8000
    inspector
      bindings
        resources
          sourcemap-typescript.ts
          sourcemap-sass.scss
        shadowdom-navigator.html
        bindings-test.js
      inspector-test.js
  (no domain)
    sourcemap-script.js
    sourcemap-style.css

Running: detachShadow2
top
  127.0.0.1:8000
    inspector
      bindings
        resources
          sourcemap-typescript.ts
        shadowdom-navigator.html
        bindings-test.js
      inspector-test.js
  (no domain)
    sourcemap-script.js
